# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VOTES-API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dashboard/management/change-password": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Cambiar la contraseña del administrador",
                "description": "Permite que un administrador cambie su contraseña, verificando la contraseña actual y asegurándose de que la nueva contraseña no sea una de las últimas tres usadas",
                "operationId": "3167d7140280964304d6b1a54d56f944",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de autorización Bearer",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer {token}"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "currentPassword",
                                    "newPassword"
                                ],
                                "properties": {
                                    "currentPassword": {
                                        "type": "string",
                                        "example": "CurrentPassword123"
                                    },
                                    "newPassword": {
                                        "type": "string",
                                        "example": "NewPassword!2024"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña cambiada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Clave actualizada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Clave actual incorrecta"
                                        },
                                        "errors": {
                                            "properties": {
                                                "currentPassword": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "La contraseña actual es requerida"
                                                    }
                                                },
                                                "newPassword": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "La nueva contraseña debe tener mínimo 6 caracteres"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/management/create-voters": {
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Crear un nuevo votante",
                "description": "Permite crear un nuevo votante (que puede ser un candidato)",
                "operationId": "df444495efbeb3cb7166f913d01634fa",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de autorización Bearer",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer {token}"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "lastName",
                                    "document",
                                    "dob",
                                    "is_candidate"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    },
                                    "is_candidate": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Votante creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Voter"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El nombre es requerido"
                                                    }
                                                },
                                                "lastName": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El apellido es requerido"
                                                    }
                                                },
                                                "document": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El documento ya está registrado"
                                                    }
                                                },
                                                "dob": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "La fecha de nacimiento es requerida"
                                                    }
                                                },
                                                "is_candidate": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo candidato o votante es requerido"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Voter": {
                "description": "Modelo de Votante",
                "properties": {
                    "id": {
                        "description": "ID del votante",
                        "type": "integer"
                    },
                    "document": {
                        "description": "Documento del votante",
                        "type": "string"
                    },
                    "name": {
                        "description": "Nombre del votante",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "Apellido del votante",
                        "type": "string"
                    },
                    "birth_day": {
                        "description": "Fecha de nacimiento del votante",
                        "type": "string",
                        "format": "date"
                    },
                    "is_candidate": {
                        "description": "Indica si el votante es candidato",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
...
}